+++ src/parkapi_sources/converters/konstanz_bike/converter.py
@@ -3,65 +3,62 @@ Copyright 2024 binary butterfly GmbH
 Use of this source code is governed by an MIT-style license that can be found in the LICENSE.txt.
 """
 
+from typing import override
+
 from validataclass.exceptions import ValidationError
-from validataclass.validators import DataclassValidator
+from validataclass.validators import AnythingValidator, DataclassValidator, ListValidator
 
-from parkapi_sources.converters.base_converter import ParkingSiteBaseConverter
-from parkapi_sources.converters.base_converter.push import CsvConverter
+from parkapi_sources.converters.base_converter.pull import ParkingSitePullConverter
 from parkapi_sources.exceptions import ImportParkingSiteException
-from parkapi_sources.models import SourceInfo, StaticParkingSiteInput
+from parkapi_sources.models import RealtimeParkingSiteInput, SourceInfo, StaticParkingSiteInput
 
-from .models import KonstanzRowInput
+from .models import KonstanzBikeParkingSiteInput
 
 
-class KonstanzBikePushConverter(CsvConverter, ParkingSiteBaseConverter):
-    konstanz_bike_row_validator = DataclassValidator(KonstanzRowInput)
-    csv_delimiter = ','
+class KonstanzBikePullConverter(ParkingSitePullConverter):
+    list_validator = ListValidator(AnythingValidator(allowed_types=[dict]))
+    konstanz_bike_validator = DataclassValidator(KonstanzBikeParkingSiteInput)
 
     source_info = SourceInfo(
         uid='konstanz_bike',
         name='Stadt Konstanz: Fahrrad-Abstellanlagen',
         public_url='https://www.konstanz.de/',
+        source_url='https://services-eu1.arcgis.com/cgMeYTGtzFtnxdsx/arcgis/rest/services/Fahrradverkehr/FeatureServer/7/query?outFields=*&where=1%3D1&f=geojson',
+        timezone='Europe/Berlin',
         has_realtime_data=False,
     )
 
-    header_mapping: dict[str, str] = {
-        'ABSTELL_NR': 'uid',
-        'Lagebezeichnung': 'address',
-        'Art': 'type',
-        'Ueberdachung': 'is_covered',
-        'Beleuchtung': 'has_lighting',
-        'Eigentuemer_Baulasttraeger': 'operator_name',
-        'Stadtteil': 'district',
-        'Kapazitaet': 'capacity',
-        'coordinates': 'coordinates',
-        'geometry': 'geometry',
-    }
+    @override
+    def get_static_parking_sites(self) -> tuple[list[StaticParkingSiteInput], list[ImportParkingSiteException]]:
+        konstanz_parking_site_inputs, static_parking_site_errors = self._get_parking_site_inputs()
 
-    def handle_csv(self, data: list[list]) -> tuple[list[StaticParkingSiteInput], list[ImportParkingSiteException]]:
-        static_parking_site_inputs: list[StaticParkingSiteInput] = []
-        static_parking_site_errors: list[ImportParkingSiteException] = []
+        static_parking_sites: list[StaticParkingSiteInput] = []
+        for parking_site_input in konstanz_parking_site_inputs:
+            static_parking_sites.append(parking_site_input.to_static_parking_site_input())
 
-        mapping: dict[str, int] = self.get_mapping_by_header(self.header_mapping, data[0])
+        return self.apply_static_patches(static_parking_sites), static_parking_site_errors
 
-        # We start at row 2, as the first one is our header
-        for row in data[1:]:
-            input_dict: dict[str, str] = {}
-            for field in self.header_mapping.values():
-                input_dict[field] = row[mapping[field]]
+    @override
+    def get_realtime_parking_sites(self) -> tuple[list[RealtimeParkingSiteInput], list[ImportParkingSiteException]]:
+        return [], []  # Only static data can be called from the API
 
+    def _get_parking_site_inputs(self) -> tuple[list[KonstanzBikeParkingSiteInput], list[ImportParkingSiteException]]:
+        konstanz_bike_parking_sites: list[KonstanzBikeParkingSiteInput] = []
+        parking_site_errors: list[ImportParkingSiteException] = []
+
+        response = self.request_get(url=self.source_info.source_url, timeout=60)
+        parking_site_inputs = response.json()['features']
+
+        for parking_site in parking_site_inputs:
             try:
-                input_data: KonstanzRowInput = self.konstanz_bike_row_validator.validate(input_dict)
+                konstanz_bike_parking_sites.append(self.konstanz_bike_validator.validate(parking_site))
             except ValidationError as e:
-                static_parking_site_errors.append(
+                parking_site_errors.append(
                     ImportParkingSiteException(
                         source_uid=self.source_info.uid,
-                        parking_site_uid=input_dict.get('id'),
-                        message=f'validation error for {input_dict}: {e.to_dict()}',
+                        parking_site_uid=parking_site.get('id'),
+                        message=f'validation error for static data {parking_site}: {e.to_dict()}',
                     ),
                 )
-                continue
 
-            static_parking_site_inputs.append(input_data.to_static_parking_site_input())
-
-        return static_parking_site_inputs, static_parking_site_errors
+        return konstanz_bike_parking_sites, parking_site_errors